Refactor BASIC runtime for numeric variants
