+++ basic/tests/expected/ld/hamurabi.out
@@ -16,7 +16,7 @@ YOU HARVESTED 3 BUSHELS PER ACRE.
 THE RATS ATE 200 BUSHELS.
 YOU NOW HAVE 2800 BUSHELS IN STORE.
 
-LAND IS TRADING AT 17 BUSHELS PER ACRE.
+LAND IS TRADING AT 20 BUSHELS PER ACRE.
 HOW MANY ACRES DO YOU WISH TO BUY HOW MANY ACRES DO YOU WISH TO SELL 
 HOW MANY BUSHELS DO YOU WISH TO FEED YOUR PEOPLE 
 HOW MANY ACRES DO YOU WISH TO PLANT WITH SEED
