+++ basic/tests/expected/ld/random.out
@@ -1,4 +1,4 @@
-0.838629710598822615
-6.80043411028139389
-0.838629710598822615
-6.80043411028139389
+3.3908526400192201
+7.82255847919924389
+3.3908526400192201
+7.82255847919924389
